Renamed /hash to /item.
Aligned /item and /entry to use the spec'd placeholders ({item-hash}, {entry-number}).

Initial attempt which generates the download archive on the fly.
It's at the wrong URL as there's an existing endpoint already there.

Removed old /download endpoint and HTML fragment.
Added functional test to check for a valid zip file response from /download endpoint.

Added Gradle build instruction to show which tests it has executed.
This allows quick check that your test ran.

Returning 406 response now instead of accidental 500 when the resource can not
 return response of type requested in accept headers.

ThrowableExceptionMapper does not handle ClientErrorException and propagates it.

Functional test confirms that the 406 is returns in case the format is not supported

Formatted the git subtree commands for clarity

Added codeblocks around the git subtree add/pull commands so they don't run into each other on a single line.
Makes it easier to read.